// src/model/UtilityLocation.ts
export class UtilityLocation {
    ID: number;
    IMAGE: string;
    WEBSITE_URL: string;
    CREATE_USER_ID: string;
    ADDRESS: string;
    EMAIL: string;
    GEO_LOCATION: string;
    PARENT_ID: number;
    PHONE: string;
    PRICE: number;
    NAME: string;
    NAME2: string;
    DESCRIPTION: string;
    LANGUAGE_ID: string;
    Distance: number;
}